+++ lib/src/extension/matrix_functions.dart
@@ -1,14 +1,5 @@
 part of matrix_utils;
 
-/* TODO:
-Matrix Functions:
-   - Exponential, logarithmic, and trigonometric functions applied element-wise
-   - Matrix rank approximation
-   - Condition number estimation
-   - Pseudoinverse calculation
-   - Matrix power (generalized, not just integer powers)
- */
-
 extension MatrixFunctions on Matrix {
   /// Element-wise sine
   Matrix sin() {
@@ -155,5 +146,115 @@ extension MatrixFunctions on Matrix {
       }
     }
     return result;
+  }
+
+  /// Element-wise exponential
+  Matrix exp() {
+    Matrix result = Matrix.zeros(rowCount, columnCount, isDouble: true);
+    for (int i = 0; i < rowCount; i++) {
+      for (int j = 0; j < columnCount; j++) {
+        result[i][j] = math.exp(this[i][j]);
+      }
+    }
+    return result;
+  }
+
+  /// Element-wise natural logarithm
+  Matrix log() {
+    Matrix result = Matrix.zeros(rowCount, columnCount, isDouble: true);
+    for (int i = 0; i < rowCount; i++) {
+      for (int j = 0; j < columnCount; j++) {
+        result[i][j] = math.log(this[i][j]);
+      }
+    }
+    return result;
+  }
+
+  /// Element-wise logarithm for any base
+  Matrix logn(num base) {
+    Matrix result = Matrix.zeros(rowCount, columnCount, isDouble: true);
+    for (int i = 0; i < rowCount; i++) {
+      for (int j = 0; j < columnCount; j++) {
+        result[i][j] = math.log(this[i][j]) / math.log(base);
+      }
+    }
+    return result;
+  }
+
+  // /// Matrix power of a positive integer value
+  // Matrix pow(int p) {
+  //   if (rowCount != columnCount) {
+  //     throw ArgumentError('Matrix must be square for power operation.');
+  //   }
+
+  //   if (p < 0) {
+  //     throw ArgumentError('Power must be non-negative.');
+  //   }
+
+  //   Matrix result = Matrix.eye(rowCount);
+  //   for (int i = 0; i < p; i++) {
+  //     result = result * this;
+  //   }
+  //   return result;
+  // }
+
+  /// Raises this matrix to a given power.
+  /// Uses eigenvalue decomposition for non-integer powers and binary exponentiation for integer powers.
+  ///
+  /// Note: For non-symmetric matrices, the eigenvalues might be complex, and thus
+  /// the result of raising them to a power could also be complex. In such cases,
+  /// the output of this function might not be accurate.
+  ///
+  /// If the matrix is not square, this method throws an ArgumentError.
+  /// If a negative eigenvalue is encountered when trying to compute a fractional power,
+  /// this method throws an ArgumentError.
+  ///
+  /// [power] - the power to which this matrix should be raised.
+  /// Returns a new matrix that is this matrix raised to the given power.
+  Matrix pow(num power) {
+    // Ensure the matrix is square
+    if (rowCount != columnCount) {
+      throw ArgumentError('Matrix must be square to raise it to a power.');
+    }
+
+    if (power < 0) {
+      throw ArgumentError('Power must be non-negative.');
+    }
+
+    if (power == power.floor()) {
+      // Use binary exponentiation for integer powers
+      Matrix result = Matrix.eye(rowCount);
+      Matrix base = this;
+      int p = power.toInt();
+
+      while (p > 0) {
+        if (p % 2 == 1) {
+          result = result * base;
+        }
+        base = base * base;
+        p = p ~/ 2;
+      }
+
+      return result;
+    } else {
+      // Perform eigenvalue decomposition
+      var eigenvalueDecomposition = decomposition.eigenvalueDecomposition();
+      Matrix D = eigenvalueDecomposition.D;
+      Matrix V = eigenvalueDecomposition.V;
+
+      // Raise eigenvalues to the given power
+      for (int i = 0; i < D.rowCount; i++) {
+        if (D[i][i] < 0 && power != power.floor()) {
+          throw ArgumentError(
+              'Negative eigenvalue found when trying to compute fractional power.');
+        }
+        D[i][i] = math.pow(D[i][i], power);
+      }
+
+      // Reconstruct the matrix using the pseudo-inverse of V
+      var result = V * D * V.pseudoInverse();
+
+      return result;
+    }
   }
 }